requirements.txt
Django==1.9
djangorestframework==3.6.3
